+++ src/components/Shared/Loading.tsx
@@ -1,3 +1,3 @@
 import * as React from "react";
 
-export default () => <img src="/assets/images/spinnyBusyIcon.gif" height="200px" width="200px"></img>;
+export default () => <img src="/assets/images/spinnyBusyIcon.gif"></img>;